requirements.txt
Flask>=3.0.0
Jinja2>=3.1.0
python-docx>=1.1.0
PyYAML>=6.0
markdown>=3.5
rapidfuzz>=3.9.0
opencv-python>=4.9.0.80
pillow>=10.0.0
loguru>=0.7.2
requests>=2.32.0
pyinstaller>=6.0.0